Iris

MLDatasets.Iris — Module

Fisher's classic iris dataset.

Measurements from 3 different species of iris: setosa, versicolor and virginica. There are 50 examples of each species.

There are 4 measurements for each example: sepal length, sepal width, petal length and petal width. The measurements are in centimeters.

The module retrieves the data from the UCI Machine Learning Repository.

NOTE: no pre-defined train-test split for this dataset, features and labels return the whole dataset.

Interface

Iris.features

Iris.labels

Utilities

Iris.download

API reference

MLDatasets.Iris.features — Function

features(; dir = nothing)

Return a 4x150 matrix containing the 4-dimensional features of each observation.

MLDatasets.Iris.labels — Function

labels(; dir = nothing)

Return a string vector of length 150 containing observations' labels.

MLDatasets.Iris.download — Function

download([dir]; [i_accept_the_terms_of_use])

Trigger the (interactive) download of the full dataset into "dir". If no dir is provided the dataset will be downloaded into "~/.julia/datadeps/Iris".

This function will display an interactive dialog unless either the keyword parameter i_accept_the_terms_of_use or the environment variable DATADEPS_ALWAY_ACCEPT is set to true. Note that using the data responsibly and respecting copyright/terms-of-use remains your responsibility.