Mutagenesis

MLDatasets.Mutagenesis — Module

Mutagenesis

Website: https://relational.fit.cvut.cz/dataset/Mutagenesis License: CC0

The Mutagenesis dataset comprises 188 molecules trialed for mutagenicity on Salmonella typhimurium, available from relational.fit.cvut.cz and CTUAvastLab/datasets.

Train, test and validation data can be loaded using following code. The withenv("DATADEPS_ALWAYS_ACCEPT"=>"true") disables the accept download prompt.

julia> using MLDatasets: Mutagenesis

julia> train_x, train_y = withenv("DATADEPS_ALWAYS_ACCEPT"=>"true") do; Mutagenesis.traindata(); end;

julia> test_x, test_y = Mutagenesis.testdata();

julia> val_x, val_y = Mutagenesis.valdata();

julia> train_x[1]
Dict{Symbol, Any} with 5 entries:
  :lumo  => -1.246
  :inda  => 0
  :logp  => 4.23
  :ind1  => 1
  :atoms => Dict{Symbol, Any}[Dict(:element=>"c", :bonds=>Dict{Symbol, Any}[Dic…

julia> train_y[1]
1