PubMed

MLDatasets.PubMed — Module

PubMed

The PubMed citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Interface

PubMed.dataset

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: Planetoid

API reference

MLDatasets.PubMed.dataset — Function

dataset(; dir=nothing)

Retrieve the PubMed dataset. The output is a named tuple with fields

julia> keys(PubMed.dataset())
(:node_features, :node_labels, :adjacency_list, :train_indices, :val_indices, :test_indices, :num_classes, :num_nodes, :num_edges, :directed)

In particular, adjacency_list is a vector of vector, where adjacency_list[i] will contain the neighbors of node i through outgoing edges.

If reverse_edges=true, the graph will contain the reverse of each edge and the graph will be undirected.

See also PubMed.

Usage Examples

using MLDatasets: PubMed
data = PubMed.dataset()
train_labels = data.node_labels[data.train_indices]