TUDataset

MLDatasets.TUDataset — Type

TUDataset(name; dir=nothing)

A variety of graph benchmark datasets, .e.g. "QM9", "IMDB-BINARY", "REDDIT-BINARY" or "PROTEINS", collected from the TU Dortmund University. Retrieve from TUDataset collection the dataset name, where name is any of the datasets available here.

A TUDataset object can be indexed to retrieve a specific graph or a subset of graphs.

Internal fields

num_nodes           # total number of nodes (considering all graphs)
num_edges           # total number of edges (considering all graphs)       
num_graphs          # total number of graphs
source              # vector of edges' source vectors      
target              # vector of edges' target vectors
graph_indicator     # graph to which a node belongs too
node_labels
edge_labels
graph_labels
node_attributes
edge_attributes
graph_attributes

See here for an in-depth description of the format.

Usage Example

using MLDatasets: TUDataset
using LightGraphs: SimpleGraph, add_edge!

data = TUDataset("PROTEINS")

# Access first graph
d1 = data[1] 

# Create a LightGraphs' graph
g = SimpleGraph(d1.num_nodes)
for (s, t) in zip(d1.source, d1.target)
    add_edge!(g, s, t)
end

# Node features
X = d1.node_attributes # (nfeatures x nnodes) matrix