Graph Datasets

A collection of datasets with an underlying graph structure. Some of these datasets contain a single graph, that can be accessed with dataset[:] or dataset[1]. Others contain many graphs, accessed through dataset[i]. Graphs are represented by the MLDatasets.Graph type.

Index

MLDatasets.CiteSeer

MLDatasets.Cora

MLDatasets.Graph

MLDatasets.KarateClub

MLDatasets.OGBDataset

MLDatasets.PolBlogs

MLDatasets.PubMed

MLDatasets.Reddit

MLDatasets.TUDataset

Documentation

MLDatasets.Graph — Type

Graph(; kws...)

A type that represents a graph and that can also store node and edge data. It doesn't distinguish between directed or undirected graph, therefore for undirected graphs will store edges in both directions. Nodes are indexed in 1:num_nodes.

Graph datasets in MLDatasets.jl contain one or more Graph objects.

Keyword Arguments

num_nodes: the number of nodes. If omitted, is inferred from edge_index.

edge_index: a tuple containing two vectors with length equal to the number of edges. The first vector contains the list of the source nodes of each edge, the second the target nodes. Defaults to (Int[], Int[]).

node_data: node-related data. Can be nothing, a named tuple of arrays or a dictionary of arrays. The arrays last dimension size should be equal to the number of nodes. Default nothing.

edge_data: edge-related data. Can be nothing, a named tuple of arrays or a dictionary of arrays. The arrays' last dimension size should be equal to the number of edges. Default nothing.

Examples

All graph datasets in MLDatasets.jl contain Graph objects:

julia> using MLDatasets: Cora

julia> d = Cora() # the Cora dataset
dataset Cora:
  metadata    =>    Dict{String, Any} with 3 entries
  graphs      =>    1-element Vector{Graph}

julia> d[1]
Graph:
  num_nodes   =>    2708
  num_edges   =>    10556
  edge_index  =>    ("10556-element Vector{Int64}", "10556-element Vector{Int64}")
  node_data   =>    (features = "1433×2708 Matrix{Float32}", targets = "2708-element Vector{Int64}", train_mask = "2708-element BitVector with 140 trues", val_mask = "2708-element BitVector with 500 trues", test_mask = "2708-element BitVector with 1000 trues")
  edge_data   =>    nothing

Let's se how to convert a Graphs.jl's graph to a MLDatasets.Graph and viceversa:

import Graphs, MLDatasets

## From Graphs.jl to MLDatasets.Graphs

# From a directed graph
g = Graphs.erdos_renyi(10, 20, is_directed=true)
s = [e.src for e in Graphs.edges(g)]
t = [e.dst for e in Graphs.edges(g)]
mlg = MLDatasets.Graph(num_nodes=10, edge_index=(s, t))

# From an undirected graph
g = Graphs.erdos_renyi(10, 20, is_directed=false)
s = [e.src for e in Graphs.edges(g)]
t = [e.dst for e in Graphs.edges(g)]
s, t = [s; t], [t; s] # adding reverse edges
mlg = MLDatasets.Graph(num_nodes=10, edge_index=(s, t))

# From MLDatasets.Graphs to Graphs.jl
s, t = mlg.edge_index
g = Graphs.DiGraph(mlg.num_nodes)
for (i, j) in zip(s, t)
    Graphs.add_edge!(g, i, j)
end

MLDatasets.CiteSeer — Type

CiteSeer(; dir=nothing)

The CiteSeer citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking

[2]: Planetoid

MLDatasets.Cora — Type

Cora()

The Cora citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. Each node has a predefined feature with 1433 dimensions. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Statistics

Nodes: 2708

Edges: 10556

Number of Classes: 7

Label split:

Train: 140

Val: 500

Test: 1000

The split is the one used in the original paper [1] and doesn't consider all nodes.

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking

[2]: [Planetoid](https://github.com/kimiyoung/planetoid

MLDatasets.KarateClub — Type

KarateClub()

The Zachary's karate club dataset originally appeared in Ref [1].

The network contains 34 nodes (members of the karate club). The nodes are connected by 78 undirected and unweighted edges. The edges indicate if the two members interacted outside the club.

The node labels indicate which community or the karate club the member belongs to. The club based labels are as per the original dataset in Ref [1]. The community labels are obtained by modularity-based clustering following Ref [2]. The data is retrieved from Ref [3] and [4]. One node per unique label is used as training data.

References

[1]: An Information Flow Model for Conflict and Fission in Small Groups

[2]: Semi-supervised Classification with Graph Convolutional Networks

[3]: PyTorch Geometric Karate Club Dataset

[4]: NetworkX Zachary's Karate Club Dataset

MLDatasets.OGBDataset — Type

OGBDataset(name; dir=nothing)

The collection of datasets from the Open Graph Benchmark: Datasets for Machine Learning on Graphs paper.

name is the name of one of the datasets (listed here) available for node prediction, edge prediction, or graph prediction tasks.

Examples

Node prediction tasks

julia> data = OGBDataset("ogbn-arxiv")
dataset OGBDataset:
  name        =>    ogbn-arxiv
  metadata    =>    Dict{String, Any} with 17 entries
  graphs      =>    1-element Vector{MLDatasets.Graph}
  graph_data  =>    nothing

julia> data[:]
Graph:
  num_nodes   =>    169343
  num_edges   =>    1166243
  edge_index  =>    ("1166243-element Vector{Int64}", "1166243-element Vector{Int64}")
  node_data   =>    (val_mask = "29799-trues BitVector", test_mask = "48603-trues BitVector", year = "169343 Vector{Int64}", features = "128×169343 Matrix{Float32}", label = "169343 Vector{Int64}", train_mask = "90941-trues BitVector")
  edge_data   =>    nothing

julia> data.metadata
Dict{String, Any} with 17 entries:
  "download_name"         => "arxiv"
  "num classes"           => 40
  "num tasks"             => 1
  "binary"                => false
  "url"                   => "http://snap.stanford.edu/ogb/data/nodeproppred/arxiv.zip"
  "additional node files" => "node_year"
  "is hetero"             => false
  "task level"            => "node"
  ⋮                       => ⋮

Edge prediction task

julia> data = OGBDataset("ogbl-collab")
dataset OGBDataset:
  name        =>    ogbl-collab
  metadata    =>    Dict{String, Any} with 15 entries
  graphs      =>    1-element Vector{MLDatasets.Graph}
  graph_data  =>    nothing

julia> data[:]
Graph:
  num_nodes   =>    235868
  num_edges   =>    2358104
  edge_index  =>    ("2358104-element Vector{Int64}", "2358104-element Vector{Int64}")
  node_data   =>    (features = "128×235868 Matrix{Float32}",)
  edge_data   =>    (year = "2×1179052 Matrix{Int64}", weight = "2×1179052 Matrix{Int64}")

Graph prediction task

julia> data = OGBDataset("ogbg-molhiv")
dataset OGBDataset:
  name        =>    ogbg-molhiv
  metadata    =>    Dict{String, Any} with 17 entries
  graphs      =>    41127-element Vector{MLDatasets.Graph}
  graph_data  =>    (labels = "41127-element Vector{Int64}", train_mask = "32901-trues BitVector", val_mask = "4113-trues BitVector", test_mask = "4113-trues BitVector")

julia> data[1]

MLDatasets.PolBlogs — Type

PolBlogs(; dir=nothing)

The Political Blogs dataset from the The Political Blogosphere and the 2004 US Election: Divided they Blog paper.

PolBlogs is a graph with 1,490 vertices (representing political blogs) and 19,025 edges (links between blogs).

The links are automatically extracted from a crawl of the front page of the blog.

Each vertex receives a label indicating the political leaning of the blog: liberal or conservative.

MLDatasets.PubMed — Type

PubMed(; dir=nothing, reverse_edges=true)

The PubMed citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking

[2]: Planetoid

MLDatasets.Reddit — Type

Reddit(; full=true, dir=nothing)

The Reddit dataset was introduced in Ref [1]. It is a graph dataset of Reddit posts made in the month of September, 2014. The dataset contains a single post-to-post graph, connecting posts if the same user comments on both. The node label in this case is one of the 41 communities, or “subreddit”s, that a post belongs to. This dataset contains 232,965 posts. The first 20 days are used for training and the remaining days for testing (with 30% used for validation). Each node is represented by a 602 word vector.

Use full=false to load only a subsample of the dataset.

References

[1]: Inductive Representation Learning on Large Graphs

[2]: Benchmarks on the Reddit Dataset

MLDatasets.TUDataset — Type

TUDataset(name; dir=nothing)

A variety of graph benchmark datasets, .e.g. "QM9", "IMDB-BINARY", "REDDIT-BINARY" or "PROTEINS", collected from the TU Dortmund University. Retrieve from the TUDataset collection the dataset name, where name is any of the datasets available here.

A TUDataset object can be indexed to retrieve a specific graph or a subset of graphs.

See here for an in-depth description of the format.

Usage Example

julia> data = TUDataset("PROTEINS")
dataset TUDataset:
  name        =>    PROTEINS
  metadata    =>    Dict{String, Any} with 1 entry
  graphs      =>    1113-element Vector{MLDatasets.Graph}
  graph_data  =>    (targets = "1113-element Vector{Int64}",)
  num_nodes   =>    43471
  num_edges   =>    162088
  num_graphs  =>    1113

julia> data[1]
(graphs = Graph(42, 162), targets = 1)