Graph Datasets

Index

MLDatasets.CiteSeer

MLDatasets.Cora

MLDatasets.KarateClub

MLDatasets.PolBlogs

MLDatasets.PubMed

MLDatasets.OGBDataset

MLDatasets.TUDataset

MLDatasets.CiteSeer.dataset

MLDatasets.Cora.dataset

MLDatasets.KarateClub.edge_index

MLDatasets.KarateClub.labels

MLDatasets.PolBlogs.edge_index

MLDatasets.PolBlogs.labels

MLDatasets.PubMed.dataset

Documentation

MLDatasets.CiteSeer — Module

CiteSeer

The CiteSeer citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Interface

CiteSeer.dataset

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: Planetoid

MLDatasets.CiteSeer.dataset — Function

dataset(; dir=nothing, reverse_edges=true)

Retrieve the CiteSeer dataset. The output is a named tuple with fields

julia> keys(CiteSeer.dataset())
(:node_features, :node_labels, :adjacency_list, :train_indices, :val_indices, :test_indices, :num_classes, :num_nodes, :num_edges, :directed)

In particular, adjacency_list is a vector of vector, where adjacency_list[i] will contain the neighbors of node i through outgoing edges.

If reverse_edges=true, the graph will contain the reverse of each edge and the graph will be undirected.

See also CiteSeer.

Usage Examples

using MLDatasets: CiteSeer
data = CiteSeer.dataset()
train_labels = data.node_labels[data.train_indices]

MLDatasets.Cora — Module

Cora

The Cora citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. Each node has a predefined feature with 1433 dimensions. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Statistics

Nodes: 2708

Edges: 10556

Number of Classes: 7

Label split:

Train: 140

Val: 500

Test: 1000

The split is the one used in the original paper [1] and doesn't consider all nodes.

Interface

Cora.dataset

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: [Planetoid](https://github.com/kimiyoung/planetoid

MLDatasets.Cora.dataset — Function

dataset(; dir=nothing, reverse_edges=true)

Retrieve the Cora dataset. The output is a named tuple with fields

julia> keys(Cora.dataset())
(:node_features, :node_labels, :adjacency_list, :train_indices, :val_indices, :test_indices, :num_classes, :num_nodes, :num_edges, :directed)

In particular, adjacency_list is a vector of vector, where adjacency_list[i] will contain the neighbors of node i through outgoing edges.

If reverse_edges=true, the graph will contain the reverse of each edge and the graph will be undirected.

See also Cora.

Usage Examples

using MLDatasets: Cora

data = Cora.dataset()
train_labels = data.node_labels[data.train_indices]

MLDatasets.OGBDataset — Type

OGBDataset(name; dir=nothing)

The collection of datasets from the Open Graph Benchmark: Datasets for Machine Learning on Graphs paper.

name is the name of one of the dasets (listed here) available for node prediction, edge prediction, or graph prediction tasks.

The OGBDataset type stores the graphs internally as dictionary objects. The key "edgeindex" contains `2 x numedges`, where the first and second column contain the source and target nodes of each edge respectively.

Examples

Node prediction tasks

julia> data = OGBDataset("ogbn-arxiv")
OGBDataset{Vector{Any}}:
  name => ogbn-arxiv
  path => /home/carlo/.julia/datadeps/OGBDataset/arxiv
  metadata => Dict{String, Any} with 15 entries
  graphs => 1-element Vector{Dict}
  labels => 1-element Vector{Any}
  split => Dict{String, Any} with 3 entries

julia> data.metadata
Dict{String, Any} with 15 entries:
  "num classes"           => 40
  "binary"                => false
  "is hetero"             => false
  "eval metric"           => "acc"
  "task type"             => "multiclass classification"
  "version"               => 1
  "split"                 => "time"
  "download_name"         => "arxiv"
  "num tasks"             => 1
  "url"                   => "http://snap.stanford.edu/ogb/data/nodeproppred/arxiv.zip"
  "additional node files" => "node_year"
  "add_inverse_edge"      => false
  "has_node_attr"         => true
  "additional edge files" => nothing
  "has_edge_attr"         => false

julia> data.split
Dict{String, Any} with 3 entries:
  "test_idx"  => [347, 399, 452, 481, 489, 491, 527, 538, 541, 603  …  169334, 169335, 169336, 169337, 169338, 169339, 169340, 169341, 169342, 169343]
  "train_idx" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  169110, 169112, 169113, 169114, 169115, 169116, 169118, 169146, 169149, 169252]
  "val_idx"   => [350, 358, 367, 383, 394, 422, 430, 436, 468, 470  …  169089, 169096, 169108, 169111, 169128, 169156, 169177, 169186, 169262, 169297]

julia> length(data)
1

julia> graph, labels = data[1];

julia> graph
Dict{String, Any} with 6 entries:
  "edge_index" => [104448 13092; 15859 47284; … ; 45119 162538; 45119 72718]
  "edge_feat"  => nothing
  "node_feat"  => Float32[-0.057943 -0.1245 … -0.138236 -0.029875; -0.05253 -0.070665 … 0.040885 0.268417; … ; -0.172796 -0.372111 … -0.041253 0.077647; -0.140059 -0.301036 … -0.376132 -0.091018]
  "num_nodes"  => 169343
  "node_year"  => [2013 2015 … 2020 2020]
  "num_edges"  => 1166243

julia> source, target = graph["edge_index][:,1], graph["edge_index][:,2];

Edge prediction task

julia> data = OGBDataset("ogbl-collab")
OGBDataset{Nothing}:
  name => ogbl-collab
  path => /home/carlo/.julia/datadeps/OGBDataset/collab
  metadata => Dict{String, Any} with 13 entries
  graphs => 1-element Vector{Dict}
  labels => nothing
  split => Dict{String, Any} with 3 entries

julia> graph = data[1]  # no labels for this dataset
Dict{String, Any} with 7 entries:
  "edge_index"  => [150990 224882; 150990 224882; … ; 221742 135759; 207233 140615]
  "edge_feat"   => nothing
  "node_feat"   => Float32[-0.177486 -0.237488 … 0.004236 -0.035025; -0.10298 0.022193 … 0.031942 -0.118059; … ; 0.003879 0.062124 … 0.05208 -0.176961; -0.276317 -0.081464 … -0.201557 -0.258715]
  "num_nodes"   => 235868
  "edge_year"   => [2004 2002 … 2006 1984; 2004 2002 … 2006 1984]
  "edge_weight" => [2 1 … 1 1; 2 1 … 1 1]
  "num_edges"   => 2358104

Graph prediction task

julia> data = OGBDataset("ogbg-molhiv")
OGBDataset{Matrix{Int64}}:
  name => ogbg-molhiv
  path => /home/carlo/.julia/datadeps/OGBDataset/molhiv
  metadata => Dict{String, Any} with 15 entries
  graphs => 41127-element Vector{Dict}
  labels => 1×41127 Matrix{Int64}
  split => Dict{String, Any} with 3 entries

julia> length(data)
41127

julia> graph, labels = data[10]
(Dict{String, Any}("edge_index" => [-202 -201; -201 -200; … ; -198 -184; -201 -202], "node_feat" => Float32[7.0 6.0 … 7.0 7.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0], "edge_feat" => Float32[0.0 0.0 … 0.0 1.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 1.0], "num_nodes" => 20, "num_edges" => 42), [0])

julia> graph, labels = data[10];

julia> graph
Dict{String, Any} with 5 entries:
  "edge_index" => [1 2; 2 3; … ; 5 19; 2 1]
  "edge_feat"  => Float32[0.0 0.0 … 0.0 1.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 1.0]
  "node_feat"  => Float32[7.0 6.0 … 7.0 7.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0]
  "num_nodes"  => 20
  "num_edges"  => 42

julia> labels
1-element Vector{Int64}:
 0

MLDatasets.PolBlogs — Module

PolBlogs

The Political Blogs dataset from the The Political Blogosphere and the 2004 US Election: Divided they Blog paper.

PolBlogs is a graph with 1,490 vertices (representing political blogs) and 19,025 edges (links between blogs).

The links are automatically extracted from a crawl of the front page of the blog.

Each vertex receives a label indicating the political leaning of the blog: liberal or conservative.

Interface

PolBlogs.edge_index

PolBlogs.labels

MLDatasets.PolBlogs.edge_index — Function

edge_index(; dir = nothing)

Returns a 19025 x 2 matrix containing edge indices where first column as source node and second column as target node together they represent an edge

using MLDatasets: PolBlogs
adj = PolBlogs.edge_index()

MLDatasets.PolBlogs.labels — Function

labels(; dir = nothing)

Returns a vector containing the 1490 labels.

using MLDatasets: PolBlogs
labels = PolBlogs.labels()

MLDatasets.PubMed — Module

PubMed

The PubMed citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Interface

PubMed.dataset

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: Planetoid

MLDatasets.PubMed.dataset — Function

dataset(; dir=nothing)

Retrieve the PubMed dataset. The output is a named tuple with fields

julia> keys(PubMed.dataset())
(:node_features, :node_labels, :adjacency_list, :train_indices, :val_indices, :test_indices, :num_classes, :num_nodes, :num_edges, :directed)

In particular, adjacency_list is a vector of vector, where adjacency_list[i] will contain the neighbors of node i through outgoing edges.

If reverse_edges=true, the graph will contain the reverse of each edge and the graph will be undirected.

See also PubMed.

Usage Examples

using MLDatasets: PubMed
data = PubMed.dataset()
train_labels = data.node_labels[data.train_indices]

MLDatasets.TUDataset — Type

TUDataset(name; dir=nothing)

A variety of graph benchmark datasets, .e.g. "QM9", "IMDB-BINARY", "REDDIT-BINARY" or "PROTEINS", collected from the TU Dortmund University. Retrieve from TUDataset collection the dataset name, where name is any of the datasets available here.

A TUDataset object can be indexed to retrieve a specific graph or a subset of graphs.

Internal fields

num_nodes           # total number of nodes (considering all graphs)
num_edges           # total number of edges (considering all graphs)       
num_graphs          # total number of graphs
source              # vector of edges' source vectors      
target              # vector of edges' target vectors
graph_indicator     # graph to which a node belongs too
node_labels
edge_labels
graph_labels
node_attributes
edge_attributes
graph_attributes

See here for an in-depth description of the format.

Usage Example

using MLDatasets: TUDataset
using LightGraphs: SimpleGraph, add_edge!

data = TUDataset("PROTEINS")

# Access first graph
d1 = data[1] 

# Create a LightGraphs' graph
g = SimpleGraph(d1.num_nodes)
for (s, t) in zip(d1.source, d1.target)
    add_edge!(g, s, t)
end

# Node features
X = d1.node_attributes # (nfeatures x nnodes) matrix

MLDatasets.KarateClub — Module

Zachary's Karate Club

The Karate Club dataset originally appeared in Ref [1].

The network contains 34 nodes (members of the karate club). The nodes are connected by 78 undirected and unweighted edges. The edges indicate if the two members interacted outside the club.

The node labels indicate which community or the karate club the member belongs to. The club based labels are as per the original dataset in Ref [1]. The community labels are obtained by modularity-based clustering following Ref [2]. The data is retrieved from Ref [3] and [4]. One node per unique label is used as training data.

Interface

KarateClub.edge_index

KarateClub.labels

References

[1]: An Information Flow Model for Conflict and Fission in Small Groups [2]: Semi-supervised Classification with Graph Convolutional Networks [3]: PyTorch Geometric Karate Club Dataset [4]: NetworkX Zachary's Karate Club Dataset

MLDatasets.KarateClub.edge_index — Function

edge_index()

Returns an adjacency list where the first and second vector contain the source and target nodes of each edge respectively.

using MLDatasets: KarateClub
adj = KarateClub.edge_index()

MLDatasets.KarateClub.labels — Function

labels(mode = :club)

Returns a vector containing the node labels indicating which club or community the node belongs to.

The type of labels can be specifed using the mode variable which takes two values: :club and :community.

using MLDatasets: KarateClub

## Club labels
club_labels = KarateClub.labels(:club)

## Community labels
com_labels = KarateClub.labels(:community)