Graph Datasets

Index

MLDatasets.CiteSeer

MLDatasets.Cora

MLDatasets.Graph

MLDatasets.KarateClub

MLDatasets.OGBDataset

MLDatasets.PolBlogs

MLDatasets.PubMed

MLDatasets.Reddit

MLDatasets.TUDataset

Documentation

MLDatasets.Graph — Type

Graph(; kws...)

A type representing a graph (a set of nodes and edges) that can also store additional node and edge data.

Graph datasets in MLDatasets.jl contain one or more Graph objects.

Fields

num_nodes

num_edges

edge_index

node_data

edge_data

MLDatasets.CiteSeer — Type

CiteSeer(; dir=nothing)

The CiteSeer citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: Planetoid

MLDatasets.Cora — Type

Cora()

The Cora citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. Each node has a predefined feature with 1433 dimensions. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Statistics

Nodes: 2708

Edges: 10556

Number of Classes: 7

Label split:

Train: 140

Val: 500

Test: 1000

The split is the one used in the original paper [1] and doesn't consider all nodes.

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: [Planetoid](https://github.com/kimiyoung/planetoid

MLDatasets.KarateClub — Type

Zachary's Karate Club

The Karate Club dataset originally appeared in Ref [1].

The network contains 34 nodes (members of the karate club). The nodes are connected by 78 undirected and unweighted edges. The edges indicate if the two members interacted outside the club.

The node labels indicate which community or the karate club the member belongs to. The club based labels are as per the original dataset in Ref [1]. The community labels are obtained by modularity-based clustering following Ref [2]. The data is retrieved from Ref [3] and [4]. One node per unique label is used as training data.

References

[1]: An Information Flow Model for Conflict and Fission in Small Groups [2]: Semi-supervised Classification with Graph Convolutional Networks [3]: PyTorch Geometric Karate Club Dataset [4]: NetworkX Zachary's Karate Club Dataset

MLDatasets.OGBDataset — Type

OGBDataset(name; dir=nothing)

The collection of datasets from the Open Graph Benchmark: Datasets for Machine Learning on Graphs paper.

name is the name of one of the dasets (listed here) available for node prediction, edge prediction, or graph prediction tasks.

The OGBDataset type stores the graphs internally as dictionary objects. The key "edgeindex" contains `2 x numedges`, where the first and second column contain the source and target nodes of each edge respectively.

Examples

Node prediction tasks

julia> data = OGBDataset("ogbn-arxiv")
OGBDataset{Vector{Any}}:
  name => ogbn-arxiv
  path => /home/carlo/.julia/datadeps/OGBDataset/arxiv
  metadata => Dict{String, Any} with 15 entries
  graphs => 1-element Vector{Dict}
  labels => 1-element Vector{Any}
  split => Dict{String, Any} with 3 entries

julia> data.metadata
Dict{String, Any} with 15 entries:
  "num classes"           => 40
  "binary"                => false
  "is hetero"             => false
  "eval metric"           => "acc"
  "task type"             => "multiclass classification"
  "version"               => 1
  "split"                 => "time"
  "download_name"         => "arxiv"
  "num tasks"             => 1
  "url"                   => "http://snap.stanford.edu/ogb/data/nodeproppred/arxiv.zip"
  "additional node files" => "node_year"
  "add_inverse_edge"      => false
  "has_node_attr"         => true
  "additional edge files" => nothing
  "has_edge_attr"         => false

julia> data.split
Dict{String, Any} with 3 entries:
  "test_idx"  => [347, 399, 452, 481, 489, 491, 527, 538, 541, 603  …  169334, 169335, 169336, 169337, 169338, 169339, 169340, 169341, 169342, 169343]
  "train_idx" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  169110, 169112, 169113, 169114, 169115, 169116, 169118, 169146, 169149, 169252]
  "val_idx"   => [350, 358, 367, 383, 394, 422, 430, 436, 468, 470  …  169089, 169096, 169108, 169111, 169128, 169156, 169177, 169186, 169262, 169297]

julia> length(data)
1

julia> graph, labels = data[1];

julia> graph
Dict{String, Any} with 6 entries:
  "edge_index" => [104448 13092; 15859 47284; … ; 45119 162538; 45119 72718]
  "edge_feat"  => nothing
  "node_feat"  => Float32[-0.057943 -0.1245 … -0.138236 -0.029875; -0.05253 -0.070665 … 0.040885 0.268417; … ; -0.172796 -0.372111 … -0.041253 0.077647; -0.140059 -0.301036 … -0.376132 -0.091018]
  "num_nodes"  => 169343
  "node_year"  => [2013 2015 … 2020 2020]
  "num_edges"  => 1166243

julia> source, target = graph["edge_index][:,1], graph["edge_index][:,2];

Edge prediction task

julia> data = OGBDataset("ogbl-collab")
OGBDataset{Nothing}:
  name => ogbl-collab
  path => /home/carlo/.julia/datadeps/OGBDataset/collab
  metadata => Dict{String, Any} with 13 entries
  graphs => 1-element Vector{Dict}
  labels => nothing
  split => Dict{String, Any} with 3 entries

julia> graph = data[1]  # no labels for this dataset
Dict{String, Any} with 7 entries:
  "edge_index"  => [150990 224882; 150990 224882; … ; 221742 135759; 207233 140615]
  "edge_feat"   => nothing
  "node_feat"   => Float32[-0.177486 -0.237488 … 0.004236 -0.035025; -0.10298 0.022193 … 0.031942 -0.118059; … ; 0.003879 0.062124 … 0.05208 -0.176961; -0.276317 -0.081464 … -0.201557 -0.258715]
  "num_nodes"   => 235868
  "edge_year"   => [2004 2002 … 2006 1984; 2004 2002 … 2006 1984]
  "edge_weight" => [2 1 … 1 1; 2 1 … 1 1]
  "num_edges"   => 2358104

Graph prediction task

julia> data = OGBDataset("ogbg-molhiv")
OGBDataset{Matrix{Int64}}:
  name => ogbg-molhiv
  path => /home/carlo/.julia/datadeps/OGBDataset/molhiv
  metadata => Dict{String, Any} with 15 entries
  graphs => 41127-element Vector{Dict}
  labels => 1×41127 Matrix{Int64}
  split => Dict{String, Any} with 3 entries

julia> length(data)
41127

julia> graph, labels = data[10]
(Dict{String, Any}("edge_index" => [-202 -201; -201 -200; … ; -198 -184; -201 -202], "node_feat" => Float32[7.0 6.0 … 7.0 7.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0], "edge_feat" => Float32[0.0 0.0 … 0.0 1.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 1.0], "num_nodes" => 20, "num_edges" => 42), [0])

julia> graph, labels = data[10];

julia> graph
Dict{String, Any} with 5 entries:
  "edge_index" => [1 2; 2 3; … ; 5 19; 2 1]
  "edge_feat"  => Float32[0.0 0.0 … 0.0 1.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 1.0]
  "node_feat"  => Float32[7.0 6.0 … 7.0 7.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0]
  "num_nodes"  => 20
  "num_edges"  => 42

julia> labels
1-element Vector{Int64}:
 0

MLDatasets.PolBlogs — Type

PolBlogs(; dir=nothing)

The Political Blogs dataset from the The Political Blogosphere and the 2004 US Election: Divided they Blog paper.

PolBlogs is a graph with 1,490 vertices (representing political blogs) and 19,025 edges (links between blogs).

The links are automatically extracted from a crawl of the front page of the blog.

Each vertex receives a label indicating the political leaning of the blog: liberal or conservative.

MLDatasets.PubMed — Type

PubMed(; dir=nothing, reverse_edges=true)

The PubMed citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: Planetoid

MLDatasets.Reddit — Type

Reddit(; full=true, dir=nothing)

The Reddit dataset was introduced in Ref [1]. It is a graph dataset of Reddit posts made in the month of September, 2014. The dataset contains a single post-to-post graph, connecting posts if the same user comments on both. The node label in this case is one of the 41 communities, or “subreddit”s, that a post belongs to. This dataset contains 232,965 posts. The first 20 days are used for training and the remaining days for testing (with 30% used for validation). Each node is represented by a 602 word vector.

Use full=false to lod only a subsample of the dataset.

References

[1]: Inductive Representation Learning on Large Graphs [2]: Benchmarks on the Reddit Dataset

MLDatasets.TUDataset — Type

TUDataset(name; dir=nothing)

A variety of graph benchmark datasets, .e.g. "QM9", "IMDB-BINARY", "REDDIT-BINARY" or "PROTEINS", collected from the TU Dortmund University. Retrieve from TUDataset collection the dataset name, where name is any of the datasets available here.

A TUDataset object can be indexed to retrieve a specific graph or a subset of graphs.

See here for an in-depth description of the format.

Usage Example

using MLDatasets: TUDataset
using LightGraphs: SimpleGraph, add_edge!

data = TUDataset("PROTEINS")

# Access first graph
d1 = data[1] 

# Create a LightGraphs' graph
g = SimpleGraph(d1.num_nodes)
for (s, t) in zip(d1.source, d1.target)
    add_edge!(g, s, t)
end

# Node features
X = d1.node_attributes # (nfeatures x nnodes) matrix